requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
databricks-sdk==0.102.0
debugpy==1.8.20
decorator==5.2.1
delta-spark==3.1.0
docker==7.1.0
executing==2.2.1
fastapi==0.135.2
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.3.2
gunicorn==25.3.0
h11==0.16.0
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==8.37.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
nest-asyncio==1.6.0
numpy==2.4.4
nvidia-nccl-cu12==2.29.7; platform_system == "Linux"
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
prettytable==3.17.0
prompt_toolkit==3.0.52
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
pyspark==3.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
skops==0.13.0
smmap==5.0.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
stack-data==0.6.3
starlette==1.0.0
threadpoolctl==3.6.0
tornado==6.5.5
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
Werkzeug==3.1.7
xgboost==3.2.0
zipp==3.23.0
feedparser>=6.0,<7.0
redis>=5.0,<6.0
kafka-python>=2.0,<3.0
google-api-python-client>=2.137,<3.0
